second commit

+++ hunter.cpp
@@ -20,7 +20,7 @@ bool isValid(int distance, const vector<int>& x, const vector<int>& y, int n, in
  
 int main() { 
    int a;
-   int b;
+   double b;
     int n, m; 
     cin >> n >> m;
